requirements.txt
langchain==0.1.9
langchain-community==0.0.25
langchain-core==0.1.28
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
langchain_openai
langgraph


